Clear other users when disconnect

+++ app.js
@@ -16,7 +16,7 @@ const s = {
     CONNECT: "connection",
     DISCONNECT: "disconnect",
     CHAT_JOIN: "chat.join",
-    CHAT_DISCONNECT: "chat.disconnect",
+    ROOMS_LEAVE: "rooms.leave",
     ROOMS_GETLIST: "rooms.getList",
     ROOMS_JOIN: "rooms.join",
     ROOMS_GETUSERS: "rooms.getUsers",
@@ -46,44 +46,43 @@ io.on(s.CONNECT, function (socket) {
 
         client.hmset(`rooms:${socket.room}:users:${socket.username}`, ['username', socket.username, 'status', socket.status]);
 
-        // 2. Broadcast username
-        socket.to(socket.room).emit(s.CHAT_JOIN, {'username': socket.username});
-
         client.smembers("rooms", (err, res) => {
             socket.emit(s.ROOMS_GETLIST, res)
-        });
-
-        client.lrange(`rooms:${socket.room}:messages`, 0, 20, (err, res) => {
-            socket.emit(s.ROOMS_GETMESSAGES, res.reverse().map(json => JSON.parse(json)));
         });
     });
 
     socket.on(s.ROOMS_JOIN, room => {
-        socket.leave(socket.room, () => {
-            consoleLog('rooms', 'join', `User ${socket.username} leave room "${socket.room}" and is joining room "${room}"`);
+        if (room !== socket.room) {
+            socket.status = "active";
 
-            client.del(`rooms:${socket.room}:users:${socket.username}`);
+            socket.leave(socket.room, () => {
+                consoleLog('rooms', 'join', `User ${socket.username} leave room "${socket.room}" and is joining room "${room}"`);
 
-            socket.join(room, () => {
-                socket.room = room;
+                io.in(socket.room).emit(s.ROOMS_LEAVE, socket.username);
 
-                client.hmset(`rooms:${socket.room}:users:${socket.username}`, ['username', socket.username, 'status', socket.status]);
+                client.del(`rooms:${socket.room}:users:${socket.username}`);
 
-                socket.to(socket.room).emit(s.ROOMS_NEWUSER, socket.username);
+                socket.join(room, () => {
+                    socket.room = room;
 
-                client.lrange(`rooms:${socket.room}:messages`, 0, 20, (err, res) => {
-                    socket.emit(s.ROOMS_GETMESSAGES, res.reverse().map(json => JSON.parse(json)));
-                });
+                    client.hmset(`rooms:${socket.room}:users:${socket.username}`, ['username', socket.username, 'status', socket.status]);
 
-                client.keys(`rooms:${socket.room}:users:*`, (err, userKeys) => {
-                    for (let userKey of userKeys) {
-                        client.hgetall(userKey, (err, user) => {
-                            socket.emit(s.ROOMS_GETUSERS, user);
-                        })
-                    }
+                    socket.to(socket.room).emit(s.ROOMS_NEWUSER, {'username': socket.username, 'status': socket.status});
+
+                    client.lrange(`rooms:${socket.room}:messages`, 0, 20, (err, res) => {
+                        socket.emit(s.ROOMS_GETMESSAGES, res.reverse().map(json => JSON.parse(json)));
+                    });
+
+                    client.keys(`rooms:${socket.room}:users:*`, (err, userKeys) => {
+                        for (let userKey of userKeys) {
+                            client.hgetall(userKey, (err, user) => {
+                                socket.emit(s.ROOMS_GETUSERS, user);
+                            });
+                        }
+                    });
                 });
             });
-        });
+        }
     });
 
     socket.on(s.ROOMS_POSTMESSAGE, (message) => {
@@ -99,11 +98,11 @@ io.on(s.CONNECT, function (socket) {
         console.log('user disconnected');
 
         if (socket.username !== undefined) {
-            socket.broadcast.emit(s.CHAT_DISCONNECT, {username: socket.username});
+            io.in(socket.room).emit(s.ROOMS_LEAVE, socket.username);
             client.del(`rooms:${socket.room}:users:${socket.username}`);
             client.del(`users:${socket.username}`);
         }
-    })
+    });
 });
 
 http.listen(3000, () => console.log('Listening on ' + 'http://localhost:3000\n'.green));
